refactor: rename Holding to Investment
Investment better describe the intent of those objects, while Holding
might be a little confusing.
